Fix bugs, improve code, ....

+++ Services/CategoryService.cs
@@ -1,8 +1,9 @@
-﻿using FastFood.DB;
+﻿using FastFood.DB.Entities;
 using FastFood.Models;
 using FastFood.Models.ViewModels;
 using FastFood.Repositories.Interfaces;
 using FastFood.Services.Interfaces;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using X.PagedList;
 using X.PagedList.Extensions;
 
@@ -59,9 +60,15 @@ namespace FastFood.Services
         public async Task<(bool, string)> DeleteCategory(int categoryId)
         {
             var category = await this._categoryRepository.GetCategoryById(categoryId);
+            var hasProducts = category.Products.Any();
 
             if (category != null)
             {
+                if (hasProducts)
+                {
+                    return (false, $"Danh mục {categoryId} đã có sản phẩm. Không thể xóa !");
+                }
+
                 CommonHelper.DeleteFile(category.BackgroundImage ?? string.Empty);
                 CommonHelper.DeleteFile(category.ThumbnailImage ?? string.Empty);
 
@@ -74,17 +81,18 @@ namespace FastFood.Services
 
         public async Task<(bool, string)> DeleteCategories(int[] categoryIds)
         {
-            int totalDeleted = 0;
+            List<int> successIds = new();
+            List<int> failedIds = new();
+
             foreach (int categoryId in categoryIds)
             {
                 (bool success, string message) = await this.DeleteCategory(categoryId);
-                if (!success)
-                {
-                    return (false, message);
-                }
-                totalDeleted++;
+                if (success)
+                    successIds.Add(categoryId);
+                else 
+                    failedIds.Add(categoryId);
             }
-            return (true, $"Đã xóa thành công {totalDeleted} danh mục !");
+            return (true, $"Đã xóa thành công {successIds.Count}/{failedIds.Count} danh mục. Các id danh mục lỗi gồm: {failedIds.ToArray()}");
         }
 
         public async Task<(bool, string)> EditCategory(EditCategoryViewModel editCategoryViewModel)
@@ -98,6 +106,19 @@ namespace FastFood.Services
                 return (true, $"Sửa danh mục {editCategoryViewModel.CategoryId} thành công !");
             }
             return (false, $"Không tìm thấy danh mục {editCategoryViewModel.CategoryId} !");
+        }
+
+        public async Task<SelectList> GetCustomCategoriesSelectList()
+        {
+            var categories = await this._categoryRepository.GetCategories();
+            return new SelectList(
+                categories
+                .Select(p => new SelectListItem
+                {
+                    Text = p.CategoryName,
+                    Value = p.CategoryId.ToString()
+                }).ToList(), "Value", "Text");
+
         }
     }
 }
